refactor: move everything related to terminal ui into src as an API
