+++ packages/skott/src/rendering/terminal/cli-config.ts
