+++ packages/skott/src/rendering/terminal/ui/console/shared.ts
