+++ packages/skott/src/rendering/terminal/ui/renderers/render-manager.ts
